\name{DPmCER-package}
\alias{DPmCER-package}
\alias{DPmCER}
\docType{package}
\title{
\packageTitle{DPmCER}
}
\description{
\packageDescription{DPmCER}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{DPmCER}

Maintainer: \packageMaintainer{DPmCER}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = DPmCER)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{DPmCER}}
%\section{Documentation Index}{\packageIndices{DPmCER}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
